Moved code out of controller / route
